# core/vm: optimize push2 opcode