@MariusVanDerWijden 2025-04-06
Did another benchmark, a full block of PUSH2 POP instructions will go from `63745478 ns` to `16348237 ns` (63ms to 16ms) so a pretty significant improvement