@jwasinger 2025-04-08
It's pretty unfortunate that we have to resort to workarounds like this because the Go compiler is too restrictive to allow disabling inline.